start implementing final executor design
