+++ pros/Cargo.toml
@@ -21,6 +21,7 @@ no_std_io = { version = "0.6.0", features = ["alloc"] }
 pros-macros = { version = "0.1.0", path = "../pros-macros" }
 futures = { version = "0.3.28", default-features = false, features = ["alloc"] }
 concurrent-queue = { version = "2.3.0", default-features = false }
+slab = { version = "0.4.9", default-features = false }
 
 [features]
 lvgl = ["pros-sys/xapi"]
