+++ pros/src/async_runtime/task.rs
@@ -1,0 +1,21 @@
+use core::ptr::NonNull;
+
+use alloc::sync::Arc;
+use slab::Slab;
+use spin::Once;
+
+pub struct Task<T: Send> {
+    returns: Arc<Slab<Once<NonNull<()>>>>,
+    return_key: usize,
+
+    _marker: core::marker::PhantomData<T>,
+}
+
+impl<T: Send> Task<T> {
+    pub fn poll(&self) -> core::task::Poll<T> {
+        match self.returns[self.return_key].poll().map(|ptr| unsafe { (ptr.as_ptr() as *mut T).read() }) {
+            Some(val) => core::task::Poll::Ready(val),
+            None => core::task::Poll::Pending
+        }
+    }
+}
