+++ pros/src/lib.rs
@@ -1,4 +1,4 @@
-#![feature(error_in_core, stdsimd)]
+#![feature(error_in_core, stdsimd, negative_impls)]
 #![cfg_attr(not(target_arch = "wasm32"), no_std)]
 
 extern crate alloc;
